Preliminary structure for library
Have a rough structure on how user inputs will be parsed. I'm not happy with the way topology is being inputted in 'FCCU_input', I think having lists of sensors and units is unnecessary and tedious for the user, so I'm going to change this in the next version. The AVEVA format didn't provide any immediate benefits so I've opted to do it this way.